Added a use-specific utility
Since the GoNoGo analyzer expects data in a certain format, it was
necessary to reorganize and rename the files from one experiment. A
script was written to do just that. A consideration for the future would
be making the GoNoGo analyzer be able to accept a wider variety of
input, however that would lead to sloppy presentation of data. E.G. in
this scenario, people had named their "Trials" things such as "woohoo
finally done :D" - For now, data must be properly formatted in order to
be processed. This will result in a cleaner output.